% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didnp-package.R
\docType{package}
\name{didnp-package}
\alias{didnp-package}
\title{didnp: A package for computing both heterogenous and average treatment effects for the treated in a model-free differences-in-differences framework.}
\description{
The \pkg{didnp} package provides simple ways
}
\section{didnpreg}{
 \code{\link{didnpreg}}
}

\section{Bugreport}{
 Any bug or suggestion can be reported using the
\code{didnp} tracker facilities at: \url{https://github.com/}
}

\author{
Oleg Badunenko \email{oleg.badunenko@brunel.ac.uk},
Daniel J. Henderson \email{djhender@cba.ua.edu},
Stefan Sperlich \email{stefan.sperlich@unige.ch}
}
