% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didnpbsctest.R
\name{didnpbsctest}
\alias{didnpbsctest}
\alias{didnpbsctest.formula}
\alias{didnpbsctest.default}
\title{Model-free Treatment Effect Estimators}
\usage{
didnpbsctest(...)

\method{didnpbsctest}{formula}(
  formula,
  data = stop("argument 'data' is missing"),
  subset,
  boot.num = 399,
  print.level = 1,
  digits = 4,
  cores = 1,
  seed = 17345168,
  ...
)

\method{didnpbsctest}{default}(
  outcome,
  regressors,
  time,
  treated,
  treatment_period,
  weights = NULL,
  boot.num = 399,
  print.level = 1,
  digits = 4,
  cores = 1,
  seed = 17345168,
  ...
)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model. The details of model specification are given under `Details'}

\item{data}{name of the data frame; must be specified if the 'formula' method is used}

\item{subset}{NULL, optional subsample of 'data'}

\item{boot.num}{an single value specifying the number of bootstrap replications.  Default is 399.}

\item{print.level}{the level of printing; larger number implies more output is printed. Default is 1. 0 suppresses all printing.}

\item{cores}{Integer specifies the number of cores to be used for parallel computation.}

\item{seed}{integer used for the random number generation for the replication purposes. Default is 17345168.}

\item{outcome}{a vector, matrix, or data frame of length \eqn{NT}. The outcome can be a continuous or dummy variable.}

\item{regressors}{a data frame with \eqn{NT} rows that contains regressors.
A data frame class is required to identify the type/class of each regressor.}

\item{time}{a vector, matrix, or data frame of length \eqn{NT} that specifies in which period \code{id} is observed.}

\item{treatment_period}{a vector, matrix, or data frame of length \eqn{NT} with zeros for the period before treatment and ones for the period of treatment and after.}

\item{weights}{NULL,}

\item{treatment}{a vector, matrix, or data frame of length \eqn{NT} with zeros for the control and ones for the treated observations.}
}
\value{
\code{didnpbsctest} returns a list containing:
\tabular{ll}{
\code{NT}
\tab Total number of observations
\cr \tab \cr
\code{esample}
\tab A vector of TRUE/FALSE values identifying observations used in estimation. Relevant for the 'formula' method but complete cases will also be checked in the matrix method
\cr \tab \cr
\code{sample1}
\tab A vector of TRUE/FALSE values identifying treated observations.
\cr \tab \cr
\code{sample11}
\tab A vector of TRUE/FALSE values identifying treated observations right after the treatement
\cr \tab \cr
\code{sample10}
\tab A vector of TRUE/FALSE values identifying treated observations just before the treatment
\cr \tab \cr
\code{sample01}
\tab A vector of TRUE/FALSE values identifying observations in control group right after the treatement
\cr \tab \cr
\code{sample00}
\tab A vector of TRUE/FALSE values identifying observations in control group just before the treatement
\cr \tab \cr
\code{n11}
\tab A number of treated observations right after the treatment
\cr \tab \cr
\code{n10}
\tab A number of treated observations just before the treatment
\cr \tab \cr
\code{n01}
\tab A number of observations in control group right after the treatment
\cr \tab \cr
\code{n00}
\tab A number of observations in control group just before the treatment
\cr \tab \cr
\code{regressor.type}
\tab A vector of length 3 with number of continuous, unordered categorical, and ordered categorical regressors.
\cr \tab \cr
\code{bw.time}
\tab Time in seconds it took to calculate bandwidths. For bandwidth type "opt" is 0.
\cr \tab \cr
\code{boot.time}
\tab Time in seconds it took to bootstrap the standard errors.
\cr \tab \cr
\code{boot.num}
\tab Number of bootstrap replications.
\cr \tab \cr
\code{bw11}
\tab Bandwidths calculated for the sample of treated right after the treatment.
\cr \tab \cr
\code{bw10}
\tab Bandwidths calculated for the sample of treated just before the treatment.
\cr \tab \cr
\code{bw01}
\tab Bandwidths calculated for the sample of of observations in control group right after the treatment.
\cr \tab \cr
\code{bw00}
\tab Bandwidths calculated for the sample of observations in control group just before the treatement
\cr
}
}
\description{
The \code{didnpbsctest} command contains tools for computing both heterogenous and average treatment effects for the treated in a model-free differences-in-differences framework.
}
\details{
The formula shell contain multiple parts separated by '|'. An example is

form1 <- y ~ x1 + x2 | id | time | treatment | treatment_period | weights

weights can be omitted if not available

form1 <- y ~ x1 + x2 | id | time | treatment | treatment_period
}
\examples{
\dontrun{
  data(DACAsub, package = "didnp")
  # will get a data frame 'DACAsub' with 330106 rows and 18 columns

  # get the subsample
  DACAsub$mysmpl <- mysmpl <-
    DACAsub$a1922==1 & !is.na(DACAsub$a1922) &
    DACAsub$htus==1 & !is.na(DACAsub$htus)

  # generate 'treatment_period'
  DACAsub$treatment_period <- ifelse(DACAsub[,"year"]>2011,1,0)

  # define formula with the weight
  form1 <- inschool ~ fem + race + var.bpl + state + age + yrimmig +
    ageimmig | inschool | year | elig | treatment_period | perwt

  # or without the weight
  form11 <- inschool ~ fem + race + var.bpl + state + age + yrimmig +
    ageimmig | inschool | year | elig | treatment_period

  ## Syntax using formula
  # suppress output
  tym1a <- didnpbsctest(
    form1,
    data = DACAsub,
    subset = mysmpl,
    bwmethod = "opt",
    boot.num = 399,
    TTb = FALSE,
    print.level = 0,
    cores = 4)

  # Print the summary
  summary(tym1a)

  ## Use CV bandwidths
  tym1aCV <- didnpbsctest(
    form1,
    data = DACAsub,
    subset = mysmpl,
    bwmethod = "CV",
    boot.num = 399,
    TTb = FALSE,
    print.level = 1,
    cores = 4)

  # Print the summary
  summary(tym1aCV)

  ## Calculate also TTb (will take longer)
  tym1bCV <- didnpbsctest(
    form1,
    data = DACAsub,
    subset = mysmpl,
    bwmethod = "CV",
    boot.num = 399,
    TTb = TRUE,
    print.level = 1,
    cores = 4)

  # Print the summary
  summary(tym1bCV)

  ## Syntax using matrices

  tym1aM <- didnpbsctest(
    outcome = DACAsub[mysmpl,"inschool"],
    regressors = DACAsub[mysmpl,c("fem", "race", "var.bpl", "state", "age", "yrimmig", "ageimmig")],
    id = DACAsub[mysmpl,"inschool"],
    time = DACAsub[mysmpl,"year"],
    treatment = DACAsub[mysmpl,"elig"],
    treatment_period = ifelse(DACAsub[mysmpl,"year"]>2011,1,0),
    weights = DACAsub[mysmpl,"perwt"],
    bwmethod = "opt",
    boot.num = 399,
    TTb = FALSE,
    print.level = 1,
    cores = 4)

  # Print the summary
  summary(tym1aM)

}

}
\references{
... (...). This. \emph{Journal of },
\bold{1}(1), 1-1
\url{https://doi.org/10.}
}
\author{
Oleg Badunenko \email{oleg.badunenko@brunel.ac.uk},

Daniel J. Henderson \email{djhender@cba.ua.edu},

Stefan Sperlich \email{stefan.sperlich@unige.ch}
}
\keyword{did}
\keyword{np}
