% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didnpplothte.R
\name{didnpplothte}
\alias{didnpplothte}
\title{Plotting Heterogenous Treatment Effects}
\usage{
didnpplothte(
  obj,
  level = 95,
  by,
  n.intervals = 10,
  over = NULL,
  xlab = "",
  ylab = "ATET",
  over.lab = "ATET over",
  over.ci.lab = "Conf.Int.",
  point_size = 3,
  line_width = 2,
  by.labels.values = NULL,
  over.labels.values = NULL,
  text_size = 17
)
}
\arguments{
\item{obj}{an object of class "didnp".}

\item{by}{categorical or continuous.}

\item{n.intervals}{number of intervals for the numeric "by" to be split into. Default is 10.}

\item{over}{necessarily categorical.}

\item{xlab}{Label for horizontal axis. Default is "".}

\item{ylab}{Label for vertical axis. Default is "ATET".}

\item{point_size}{of ATET. Default is 3.}

\item{line_width}{of ATET for the numeric "by". Default is 2.}

\item{text_size}{for ggplot object. Default is 17.}

\item{labels.values}{for the categorical "by", the dataframe with two columns. The first column contains unique values of the "by", while the second column contains corresponding values that would appear on a graph. Default is NULL, implying that the unique values of the "by" will be used.}
}
\value{
\code{didnpplot} returns a dataframe containing containing the following columns:
\tabular{ll}{
\code{plot.a}
\tab ggplot object
\cr \tab \cr
\code{data.a}
\tab data used to produce \code{plot.a}
\cr \tab \cr
\code{plot.b}
\tab ggplot object for TTb if \code{TTb = TRUE} was used in running \link{didnpreg}
\cr \tab \cr
\code{data.a}
\tab data used to produce \code{plot.b}
\cr
}
}
\description{
The \code{didnpplothte} take output created by \link{didnpreg} and produces plots of Heterogeneous Treatment Effects with their corresponding confidence bounds. Average effects per group are produced for discrete covariates and a continuum of effects are produced for continuous covariates.
}
\details{
Vector "by" (and "over") must be of length of TTa.i if \code{TTb = FALSE} was used in running \link{didnpreg} and of length of TTb.i if \code{TTb = TRUE} was used in running \link{didnpreg}
}
\examples{
\dontrun{
  data(DACAsub, package = "didnp")
  # will get a data frame 'DACAsub' with 330106 rows and 18 columns

  # get the subsample
  DACAsub$mysmpl <- mysmpl <-
    DACAsub$a1922==1 & !is.na(DACAsub$a1922) &
    DACAsub$htus==1 & !is.na(DACAsub$htus)

  # generate 'treatment_period'
  DACAsub$treatment_period <- ifelse(DACAsub[,"year"]>2011,1,0)

  # define formula with the weight
  form1 <- inschool ~ fem + race + var.bpl + state + age + yrimmig +
    ageimmig | inschool | year | elig | treatment_period | perwt

  tym1a <- didnpreg(
    form1,
    data = DACAsub,
    subset = mysmpl,
    bwmethod = "opt",
    boot.num = 399,
    TTb = FALSE,
    print.level = 1,
    cores = 4)

  # Print the summary
  summary(tym1a)

}

}
\references{
... (...). This. \emph{Journal of },
\bold{1}(1), 1-1
\url{https://doi.org/10.}
}
\author{
Oleg Badunenko \email{oleg.badunenko@brunel.ac.uk},

Daniel J. Henderson \email{djhender@cba.ua.edu},

Stefan Sperlich \email{stefan.sperlich@unige.ch}
}
\keyword{did}
\keyword{np}
\keyword{plot}
