% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didnpreg.R
\name{didnpreg}
\alias{didnpreg}
\alias{didnpreg.formula}
\alias{didnpreg.default}
\title{Treatment Effect Estimators}
\usage{
didnpreg(...)

\method{didnpreg}{formula}(
  formula,
  data = stop("argument 'data' is missing"),
  subset,
  bwmethod = "opt",
  boot.num = 399,
  TTx = FALSE,
  TTb = TRUE,
  print.level = 1,
  cores = 4,
  seed = 17345168,
  ...
)

\method{didnpreg}{default}(
  outcome,
  regressors,
  id,
  time,
  treatment,
  treatment_period,
  weights = NULL,
  bwmethod = "opt",
  boot.num = 399,
  TTx = FALSE,
  TTb = TRUE,
  print.level = 1,
  cores = 4,
  seed = 17345168,
  ...
)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model. The details of model specification are given under `Details'}

\item{data}{name of the data frame; must be specified if the 'formula' method is used}

\item{subset}{NULL, optional subsample of 'data'}

\item{bwmethod}{"opt", # plug-in is rule of thumb for continuous and basic for categorical, can be cross-validation}

\item{boot.num}{399,}

\item{TTx}{FALSE}

\item{TTb}{FALSE}

\item{print.level}{the level of printing; larger number implies more output is printed. Default is 1. 0 suppresses all printing.}

\item{cores}{Integer specifies the number of cores to be used for parallel computation.}

\item{outcome}{a vector, matrix, or data frame of length \eqn{NT}. The outcome can be continuous or dummy.}

\item{regressors}{a data frame with \eqn{NT} rows that contains regressors.
A data frame is required to identify the type/class of each regressor.}

\item{id}{a vector, matrix, or data frame of length \eqn{NT} that identifies the unit of observation.}

\item{time}{a vector, matrix, or data frame of length \eqn{NT} that specifies in which period \code{id} is observed.}

\item{treatment}{a vector, matrix, or data frame of length \eqn{NT} with zeros for the control and ones for the treated observations.}

\item{treatment_period}{a vector, matrix, or data frame of length \eqn{NT} with zeros for the period before treatment and ones for the period of treatment and after.}

\item{weights}{NULL,}
}
\value{
\code{didnpreg} returns a list containing:
\tabular{ll}{
\code{esample}
\tab A vector of TRUE/FALSE values identifying observations used in estimation. Relevant for the 'formula' method but complete cases will also be checked in the matrix method
\cr \tab \cr
\code{esample_treated}
\tab A vector of TRUE/FALSE values identifying observations used in estimation and is treated (intersection of \code{esample} and \code{treatment})
\cr \tab \cr
\code{esample_TTa}
\tab A vector of TRUE/FALSE values identifying observations used in estimation of TTa)
\cr \tab \cr
\code{esample_TTb}
\tab A vector of TRUE/FALSE values identifying observations used in estimation of TTb. Returned only if \code{TTb = TRUE})
\cr \tab \cr
\code{N}
\tab number of units of obs, defined by the \code{id}
\cr \tab \cr
\code{N_dt}
\tab the number of observations used as training data for each conditional expectation
\cr \tab \cr
\code{T_i}
\tab number of observations in each \code{id}
\cr \tab \cr
\code{bws}
\tab bandwidth(s) for the data
\cr \tab \cr
\code{TTx}
\tab the DiD estimators of the conditional TT (also known as CATET) (equation 2.12)
\cr \tab \cr
\code{TTa}
\tab
\cr \tab \cr
\code{TTb}
\tab NULL is option TTaOnly is used
\cr
}
}
\description{
The \code{didnpreg} command contains tools for computing both heterogenous and average treatment effects for the treated in a model-free differences-in-differences framework.
}
\details{
All estimations are based. y ~ x1 + x2 | id | time | treatment | treatment_period | weights
}
\examples{
\dontrun{
  didnpreg(jhklangkjhakgkjh)
}

}
\references{
... (...). This. \emph{Journal of },
\bold{1}(1), 1-1
\url{https://doi.org/10.}
}
\author{
Oleg Badunenko \email{oleg.badunenko@brunel.ac.uk},
Daniel J. Henderson \email{djhender@cba.ua.edu},
Stefan Sperlich \email{stefan.sperlich@unige.ch}
}
\keyword{did}
\keyword{np}
