% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.didnp.R
\name{summary.didnp}
\alias{summary.didnp}
\title{Summary of objects of class "didnp"}
\usage{
\method{summary}{didnp}(obj, ...)
}
\description{
The \code{summary} prints the summary from objects of class "didnp"
}
