% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didnpreg.R
\name{didnpreg}
\alias{didnpreg}
\alias{didnpreg.formula}
\alias{didnpreg.default}
\title{Model-free Treatment Effect Estimators}
\usage{
didnpreg(...)

\method{didnpreg}{formula}(
  formula,
  data = stop("argument 'data' is missing"),
  subset,
  bws = NULL,
  bwmethod = "opt",
  boot.num = 399,
  TTx = "TTa",
  level = 95,
  print.level = 1,
  digits = 4,
  cores = 1,
  seed = 17345168,
  ...
)

\method{didnpreg}{default}(
  outcome,
  regressors,
  time,
  treated,
  treatment_period,
  weights = NULL,
  bws = NULL,
  bwmethod = "opt",
  boot.num = 399,
  TTx = "TTa",
  level = 95,
  print.level = 1,
  digits = 4,
  cores = 1,
  seed = 17345168,
  ...
)
}
\arguments{
\item{formula}{an object of class formula (or one that can be coerced to that class): a symbolic description of the model. The details of model specification are given under `Details'}

\item{data}{name of the data frame; must be specified if the 'formula' method is used}

\item{subset}{NULL, optional subsample of 'data'}

\item{bws}{a bandwidth specification. A vector of bandwidths of length corresponding to the number of regressors.}

\item{bwmethod}{bandwidth type. 2 options can be specified. "opt" is the default option, the  plug-in is rule of thumb for continuous and basic for categorical. "CV" will calculate cross-validated bandwidths.}

\item{boot.num}{an single value specifying the number of bootstrap replications.  Default is 399.}

\item{TTx}{Can take values 'TTa' or 'TTb'. 'TTa' is for estimating the Treatment Effect on the Treated by averaging over treated \emph{after} the treatment. 'TTb' is the Unconditional Treatment Effect on the Treated. TTb estimates by averaging over \emph{all} (before and after) treated. Depending on the sample, calcularing TTb may take some time. Default is 'TTb'.}

\item{print.level}{the level of printing; larger number implies more output is printed. Default is 1. 0 suppresses all printing.}

\item{cores}{Integer specifies the number of cores to be used for parallel computation.}

\item{seed}{integer used for the random number generation for the replication purposes. Default is 17345168.}

\item{outcome}{a vector, matrix, or data frame of length \eqn{NT}. The outcome can be a continuous or dummy variable.}

\item{regressors}{a data frame with \eqn{NT} rows that contains regressors.
A data frame class is required to identify the type/class of each regressor.}

\item{time}{a vector, matrix, or data frame of length \eqn{NT} that specifies in which period \code{id} is observed.}

\item{treated}{a vector, matrix, or data frame of length \eqn{NT} with zeros for the control and ones for the treated observations.}

\item{treatment_period}{a vector, matrix, or data frame of length \eqn{NT} with zeros for the period before treatment and ones for the period of treatment and after.}

\item{weights}{NULL,}
}
\value{
\code{didnpreg} returns a list containing:
\tabular{ll}{
\code{NT}
\tab Total number of observations
\cr \tab \cr
\code{esample}
\tab A vector of TRUE/FALSE values identifying observations used in estimation. Relevant for the 'formula' method but complete cases will also be checked in the matrix method
\cr \tab \cr
\code{sample1}
\tab A vector of TRUE/FALSE values identifying treated observations.
\cr \tab \cr
\code{sample11}
\tab A vector of TRUE/FALSE values identifying treated observations right after the treatment
\cr \tab \cr
\code{sample10}
\tab A vector of TRUE/FALSE values identifying treated observations just before the treatment
\cr \tab \cr
\code{sample01}
\tab A vector of TRUE/FALSE values identifying observations in control group right after the treatment
\cr \tab \cr
\code{sample00}
\tab A vector of TRUE/FALSE values identifying observations in control group just before the treatment
\cr \tab \cr
\code{n11}
\tab A number of treated observations right after the treatment
\cr \tab \cr
\code{n10}
\tab A number of treated observations just before the treatment
\cr \tab \cr
\code{n01}
\tab A number of observations in control group right after the treatment
\cr \tab \cr
\code{n00}
\tab A number of observations in control group just before the treatment
\cr \tab \cr
\code{regressor.type}
\tab A vector of length 3 with number of continuous, unordered categorical, and ordered categorical regressors.
\cr \tab \cr
\code{bwmethod}
\tab bandwidth type
\cr \tab \cr
\code{bw.time}
\tab Time in seconds it took to calculate bandwidths. For bandwidth type "opt" is 0.
\cr \tab \cr
\code{bws}
\tab Data frame with variable names, type of the regressor and bandwidths.
\cr \tab \cr
\code{boot.time}
\tab Time in seconds it took to bootstrap the standard errors.
\cr \tab \cr
\code{boot.num}
\tab Number of bootstrap replications.
\cr \tab \cr
\code{bw11}
\tab Bandwidths calculated for the sample of treated right after the treatment.
\cr \tab \cr
\code{bw10}
\tab Bandwidths calculated for the sample of treated just before the treatment.
\cr \tab \cr
\code{bw01}
\tab Bandwidths calculated for the sample of of observations in control group right after the treatment.
\cr \tab \cr
\code{bw00}
\tab Bandwidths calculated for the sample of observations in control group just before the treatment
\cr \tab \cr
\code{do.TTb}
\tab TRUE/FALSE whether to perform TTb
\cr \tab \cr
\code{TTa.positions.in.TTb}
\tab Positions of TTa observations in TTb. Only if \code{do.TTb}
\cr \tab \cr
\code{TTa}
\tab the DiD estimator of the avarage unconditional TT
\cr \tab \cr
\code{TTa.i}
\tab the DiD estimators of the unconditional TT
\cr \tab \cr
\code{TTb}
\tab the DiD estimator of the avarage unconditional TT
\cr \tab \cr
\code{TTb.i}
\tab the DiD estimators of the unconditional TT
\cr \tab \cr
\code{TTa.se}
\tab the standard error of the DiD estimator of the avarage unconditional TT
\cr \tab \cr
\code{TTb.se}
\tab the standard error of the DiD estimator of the avarage unconditional TT
\cr \tab \cr
\code{TTx}
\tab the DiD estimators of the conditional TT (also known as CATET)
\cr \tab \cr
\code{TTa.i.boot}
\tab Matrix of the size \eqn{n_{11} \times boot.num}
\cr \tab \cr
\code{TTb.i.boot}
\tab Matrix of the size \eqn{n_{1} \times boot.num}
\cr
}
}
\description{
The \code{didnpreg} command contains tools for computing both heterogenous and average treatment effects for the treated in a model-free differences-in-differences framework.
}
\details{
The formula shell contain multiple parts separated by '|'. An example is

form1 <- y ~ x1 + x2 | time | treated | treatment_period | weights

weights can be omitted if not available

form1 <- y ~ x1 + x2 | time | treated | treatment_period
}
\examples{
\dontrun{
  data(DACAsub, package = "didnp")
  # will get a data frame 'DACAsub' with 330106 rows and 18 columns

  # get the subsample
  DACAsub$mysmpl <- mysmpl <-
    DACAsub$a1922==1 & !is.na(DACAsub$a1922) &
    DACAsub$htus==1 & !is.na(DACAsub$htus)

  # generate 'treatment_period'
  DACAsub$treatment_period <- ifelse(DACAsub[,"year"]>2011,1,0)

  # define formula with the weight
  form1 <- inschool ~ fem + race + var.bpl + state + age + yrimmig +
    ageimmig | inschool | year | elig | treatment_period | perwt

  # or without the weight
  form11 <- inschool ~ fem + race + var.bpl + state + age + yrimmig +
    ageimmig | inschool | year | elig | treatment_period

  ## Syntax using formula
  # suppress output
  tym1a <- didnpreg(
    form1,
    data = DACAsub,
    subset = mysmpl,
    bwmethod = "opt",
    boot.num = 399,
    TTb = FALSE,
    print.level = 0,
    cores = 4)

  # Print the summary
  summary(tym1a)

  ## Use CV bandwidths
  tym1aCV <- didnpreg(
    form1,
    data = DACAsub,
    subset = mysmpl,
    bwmethod = "CV",
    boot.num = 399,
    TTb = FALSE,
    print.level = 1,
    cores = 4)

  # Print the summary
  summary(tym1aCV)

  ## Calculate also TTb (will take longer)
  tym1bCV <- didnpreg(
    form1,
    data = DACAsub,
    subset = mysmpl,
    bwmethod = "CV",
    boot.num = 399,
    TTb = TRUE,
    print.level = 1,
    cores = 4)

  # Print the summary
  summary(tym1bCV)

  ## Syntax using matrices

  tym1aM <- didnpreg(
    outcome = DACAsub[mysmpl,"inschool"],
    regressors = DACAsub[mysmpl,c("fem", "race", "var.bpl", "state", "age", "yrimmig", "ageimmig")],
    id = DACAsub[mysmpl,"inschool"],
    time = DACAsub[mysmpl,"year"],
    treated = DACAsub[mysmpl,"elig"],
    treatment_period = ifelse(DACAsub[mysmpl,"year"]>2011,1,0),
    weights = DACAsub[mysmpl,"perwt"],
    bwmethod = "opt",
    boot.num = 399,
    TTb = FALSE,
    print.level = 1,
    cores = 4)

  # Print the summary
  summary(tym1aM)

}

}
\references{
Daniel J. Henderson and Stefan Sperlich (2023). A Complete Framework for Model-Free Difference-in-Differences Estimation. \emph{Foundations and Trends in Econometrics},
\bold{12}(3), 232-323
\url{http://dx.doi.org/10.1561/0800000046.}
}
\author{
Oleg Badunenko \email{oleg.badunenko@brunel.ac.uk},

Daniel J. Henderson \email{djhender@cba.ua.edu},

Stefan Sperlich \email{stefan.sperlich@unige.ch}
}
\keyword{did}
\keyword{np}
