% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Unemp}
\alias{Unemp}
\title{Two Subset of the dataset about the Deferred Action for Childhood Arrivals (DACA) program}
\format{
A data frame with 17379 rows and 11 columns:
\describe{
\item{FT}{2}
\item{unemp}{2}
\item{sex}{Female}
\item{statefip}{State}
\item{school}{Post treatment period: a dummy variable that equals 1 on or after 2012}
\item{age}{Age}
\item{yrimmig}{Year of immigration}
\item{educ}{Years of education}
\item{period}{1-8}
\item{treatment_period}{Birth region (Birthplace)}
\item{eligible}{Eligible for treatment}
}
}
\usage{
Unemp
}
\description{
Two A dataset about Deferred Action for Childhood Arrivals (DACA) program containing 330106 observations
}
\keyword{datasets}
